Fixed issue with NHibernate auto mapping.
For some reason, I had to tell the auto mapping configuration to ignore
enums from being mapped. I have NO IDEA why I've had to do this, as I have
the enum convention setup exactly the way it's used in Alea, and also the
way which you're advised to use it on stack overflow.

Cheers for wasting an hour of my time, fluent NHibernate!