Consignments.
Had to end up refactoring the item history stuff so that it used its own
service. This is because an instance of the actual service is going to be
needed inside the consignment service so that there isn't a bunch of
duplicate code for creating them item history that will come about as the
result of creating a consignment. This will be done all over the place so
it makes sense that it's in its own service.