Invoicing.
Added a CreateFromPending method to the invoice item service that will
create an invoice assuming that all the values for the prices and
everything else has been validated when the pending item was created.
There are only a couple of tests for it. This commit also contains the
work I started on creating invoices from pending items.