Use latest quote item for price
This applies when there's a job item that has appeared on more than one
quote. The latest item is picked, but sadly, it's based on the date of
the quote, rather than the quote item. This is because the quote item
doesn't have a date created (and for the life of me, I don't know why I
didn't add this as a property of the quote item).